requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
transformers
torch
sentence-transformers
ollama
spacy
nltk
numpy
pandas
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

scikit-learn
textstat
langdetect
wordcloud
matplotlib
seaborn